+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/plans/PlanTest.scala
@@ -18,10 +18,9 @@
 package org.apache.spark.sql.catalyst.plans
 
 import org.scalactic.source
-import org.scalatest.Suite
 import org.scalatest.Tag
 
-import org.apache.spark.SparkFunSuite
+import org.apache.spark.{SparkFunSuite, SparkTestSuite}
 import org.apache.spark.sql.catalyst.analysis.SimpleAnalyzer
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.expressions.CodegenObjectFactoryMode
@@ -61,7 +60,7 @@ trait CodegenInterpretedPlanTest extends PlanTest {
  * Provides helper methods for comparing plans, but without the overhead of
  * mandating a FunSuite.
  */
-trait PlanTestBase extends PredicateHelper with SQLHelper { self: Suite =>
+trait PlanTestBase extends PredicateHelper with SQLHelper with SparkTestSuite {
 
   protected def normalizeExprIds(plan: LogicalPlan): LogicalPlan =
     NormalizePlan.normalizeExprIds(plan)
