+++ sql/core/src/test/scala/org/apache/spark/sql/QueryTest.scala
@@ -30,14 +30,12 @@ import org.apache.spark.sql.catalyst.util._
 import org.apache.spark.sql.classic.ClassicConversions._
 import org.apache.spark.sql.execution.{QueryExecution, SQLExecution}
 import org.apache.spark.sql.execution.columnar.InMemoryRelation
+import org.apache.spark.sql.test.SparkSessionProvider
 import org.apache.spark.sql.util.QueryExecutionListener
 import org.apache.spark.storage.StorageLevel
 import org.apache.spark.util.ArrayImplicits._
 
-
-abstract class QueryTest extends PlanTest {
-
-  protected def spark: SparkSession
+abstract class QueryTest extends PlanTest with SparkSessionProvider {
 
   /**
    * Runs the plan and makes sure the answer contains all of the keywords.
