+++ sql/core/src/test/scala/org/apache/spark/sql/execution/SparkPlanTest.scala
@@ -20,19 +20,17 @@ package org.apache.spark.sql.execution
 import scala.util.control.NonFatal
 
 import org.apache.spark.SparkFunSuite
-import org.apache.spark.sql.{DataFrame, Row, SparkSession, SQLContext}
+import org.apache.spark.sql.{DataFrame, Row, SQLContext}
 import org.apache.spark.sql.catalyst.analysis.UnresolvedAttribute
 import org.apache.spark.sql.catalyst.plans.logical.LocalRelation
 import org.apache.spark.sql.classic.ClassicConversions._
-import org.apache.spark.sql.test.SQLTestUtils
+import org.apache.spark.sql.test.{SparkSessionProvider, SQLTestUtils}
 
 /**
  * Base class for writing tests for individual physical operators. For an example of how this
  * class's test helper methods can be used, see [[SortSuite]].
  */
-private[sql] abstract class SparkPlanTest extends SparkFunSuite {
-  protected def spark: SparkSession
-
+private[sql] abstract class SparkPlanTest extends SparkFunSuite with SparkSessionProvider {
   /**
    * Runs the plan and makes sure the answer matches the expected result.
    * @param input the input data to be used.
